packages/kamiyo-hyperliquid/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn==0.30.0
slowapi==0.1.9
pydantic==2.9.0

# HTTP & Utilities
requests==2.32.3
httpx==0.27.0
python-dateutil==2.9.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# WebSocket
websockets==12.0

# Machine Learning
scikit-learn>=1.3.2
statsmodels>=0.14.1
pandas>=2.0.0,<2.1  # Python 3.8 compatible
numpy>=1.24.0,<2.0
joblib>=1.3.2

# Data Visualization (optional, for analysis)
matplotlib>=3.7.0
